MSRZ*: linux: fix issue with fast Ethernet not woking in U-Boot
In U-Boot, Ethernet does work with Fast but with Gigabit Ethernet.
It works fine in Linux. The primary problem is, that U-Boot's ravb
network driver can not handle properly two interfaces. There is
only one instance of the MII bitbanging device. As a result, the
MII access to the PHY does not work any longer, even for the first
device. For the time being we fix it by removing (uncommenting)
the secoond ethernet interface (eth1) in the DTS file. We also
overtake the phy settings from Linux, RGMII mode "rgmii-id" and
the clocks. The fix is in the "renesas-u-boot-cip" repo.

So far, the reset of the Ethernet PHYs has been released in U-Boot.
Because now only the first network device is used in U-Boot, we
define the reset-gpios in Linux's device trees as well.