MSRZG3S: add support for ARIES MSRZG3S OSM on OSMEVK or G3Sberry
The MSRZG3S is the System-In-Package from ARIES Embedded based on
Renesas RZ/G3S Family architecture. The MSRZG3S OSMs are available
as MSRZG3S-BAA with 1 GB of LPDDR4 and eMMC and MSRZG3S-A0A with
just 512 MB and without eMMC. These can be mounted on the OSMEVK
or the G3Sberry board. The support is derived from the Renesas
"smarc-rzg3s" board, which is quite similar.

We alsobump into version BSP-3.0.6-update4 for the MSRZG3S support.

Signed-off-by: Wolfgang Grandegger <wg@aries-embedded.de>