MSRZ*: remove the name fiveberry for the msrz boards, cleanup
The MSRZfive is no longer supported by ARIES Embedded. We keep
the support, but we rename and replace the name fiveberry for
the other MSRZ boards. We remove "recipes-msrz" and move the
the recipes from there to the usual places. We also remove it
from the README.md. As we are at it, we also update the
README.md, e.g. it now show properly the supported MSRZ modues
and boards and the boot mode setting.

Signed-off-by: Wolfgang Grandegger <wg@aries-embedded.de>