MSRZG3S: 2MB secure DRAM for Trust Zone and OP-TEE
In order to use Trust Zone and OP-TEE in your system, the secure area
must be hidden from Linux. By default, the BSP allocates 128MB of
space at the beginning of DDR Memory. For further information see
https://confluence.renesas.com/display/REN/RZ+BSP+Porting+-+Memory+Map.

If the build config "MSRZ_SECURE_DRAM_2MB" is set to "1" in your
"local.conf", the space will be reduced to 2MB, which is the minimum.
Linux has then 126 MB more of the memory available. It will apply
patches to Trusted-Firmware-A, U-Boot and Linux.

Signed-off-by: Wolfgang Grandegger <wg@grandegger.com>